completely rework dmg_ppu.c
dmg_ppu.c was incredibly slow, and had glaring inaccuracies, not
even passing the most basic graphical test suites. with this
rewrite it is up to a few hundred times faster, depending on
the on-screen content, and passes dmg_acid2.gb by matt currie. a
few additional rendering bugs not caught by dmg_acid2 have also
been fixed.

with this rewrite, dmg_ppu.c is no longer the slowest code in
the entire emulator, and the cpu emulator takes up much more
of the execution time, as should be the case. there is still
room for some speedups, and a lingering graphics bug in zelda:
link's awakening (however, this may not be an issue in this
module).